Refactor App component into reusable HOCs
