+++ src/index.jsx
@@ -1,79 +1,16 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import {Provider} from 'react-redux';
-import {createStore, applyMiddleware, compose} from 'redux';
-import throttle from 'redux-throttle';
-import {intlInitialState, IntlProvider} from './reducers/intl.js';
 
+import AppStateHOC from './lib/app-state-hoc.jsx';
 import GUI from './containers/gui.jsx';
-import log from './lib/log';
-import ProjectLoader from './lib/project-loader';
-import reducer from './reducers/gui';
+import ProjectLoaderHOC from './lib/project-loader-hoc.jsx';
 
 import styles from './index.css';
 
-class App extends React.Component {
-    constructor (props) {
-        super(props);
-        this.fetchProjectId = this.fetchProjectId.bind(this);
-        this.updateProject = this.updateProject.bind(this);
-        this.state = {
-            projectId: null,
-            projectData: this.fetchProjectId().length ? null : JSON.stringify(ProjectLoader.DEFAULT_PROJECT_DATA)
-        };
-    }
-    componentDidMount () {
-        window.addEventListener('hashchange', this.updateProject);
-        this.updateProject();
-    }
-    componentWillUnmount () {
-        window.removeEventListener('hashchange', this.updateProject);
-    }
-    fetchProjectId () {
-        return window.location.hash.substring(1);
-    }
-    updateProject () {
-        const projectId = this.fetchProjectId();
-        if (projectId !== this.state.projectId) {
-            if (projectId.length < 1) {
-                return this.setState({
-                    projectId: projectId,
-                    projectData: JSON.stringify(ProjectLoader.DEFAULT_PROJECT_DATA)
-                });
-            }
-            ProjectLoader.load(projectId, (err, body) => {
-                if (err) return log.error(err);
-                this.setState({projectData: body});
-            });
-            this.setState({projectId: projectId});
-        }
-    }
-    render () {
-        if (this.state.projectData === null) return null;
-        return (
-            <GUI
-                projectData={this.state.projectData}
-            />
-        );
-    }
-}
+const App = AppStateHOC(ProjectLoaderHOC(GUI));
 
 const appTarget = document.createElement('div');
 appTarget.className = styles.app;
 document.body.appendChild(appTarget);
 
-const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
-const enhancer = composeEnhancers(
-    applyMiddleware(
-        throttle(300, {leading: true, trailing: true})
-    )
-);
-const store = createStore(reducer, intlInitialState, enhancer);
-
-ReactDOM.render((
-    <Provider store={store}>
-        <IntlProvider>
-            <App />
-        </IntlProvider>
-    </Provider>
-), appTarget);
+ReactDOM.render(<App />, appTarget);
